Projeto ChatBot/requirements.txt
# Core dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.1
colorama>=0.4.6
rich>=13.7.1

# Web Interface
streamlit>=1.32.0

# Data persistence
# sqlite3 (built-in)
# json (built-in)

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
